+++ crates/cfxcore/core/src/consensus/consensus_graph/best_info_provider.rs
@@ -1,88 +1,13 @@
 use super::ConsensusGraph;
 
-pub use crate::consensus::{
-    consensus_inner::{ConsensusGraphInner, ConsensusInnerConfig},
-    consensus_trait::SharedConsensusGraph,
-};
-use crate::{
-    block_data_manager::{
-        BlockDataManager, BlockExecutionResultWithEpoch, DataVersionTuple,
-    },
-    consensus::{
-        consensus_inner::{
-            consensus_executor::ConsensusExecutionConfiguration, StateBlameInfo,
-        },
-        pos_handler::PosVerifier,
-    },
-    errors::{invalid_params, invalid_params_check, Result as CoreResult},
-    pow::{PowComputer, ProofOfWorkConfig},
-    statistics::SharedStatistics,
-    transaction_pool::SharedTransactionPool,
-    verification::VerificationConfig,
-    NodeType, Notifications,
-};
-use cfx_execute_helper::{
-    estimation::{EstimateExt, EstimateRequest},
-    exec_tracer::{
-        recover_phantom_traces, ActionType, BlockExecTraces, LocalizedTrace,
-        TraceFilter,
-    },
-    phantom_tx::build_bloom_and_recover_phantom,
-};
-use cfx_executor::{
-    executive::ExecutionOutcome, spec::CommonParams, state::State,
-};
-use cfx_rpc_eth_types::EvmOverrides;
-use geth_tracer::GethTraceWithHash;
-
-use alloy_rpc_types_trace::geth::GethDebugTracingOptions;
-use cfx_internal_common::ChainIdParams;
-use cfx_parameters::{
-    consensus::*,
-    consensus_internal::REWARD_EPOCH_COUNT,
-    rpc::{
-        GAS_PRICE_BLOCK_SAMPLE_SIZE, GAS_PRICE_DEFAULT_VALUE,
-        GAS_PRICE_TRANSACTION_SAMPLE_SIZE,
-    },
-};
-use cfx_rpc_cfx_types::PhantomBlock;
-use cfx_statedb::StateDb;
-use cfx_storage::{
-    state::StateTrait, state_manager::StateManagerTrait, StorageState,
-};
-use cfx_types::{AddressWithSpace, AllChainID, Bloom, Space, H256, U256};
-use either::Either;
-use itertools::Itertools;
-use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
+use cfx_types::{AllChainID, H256, U256};
 use malloc_size_of_derive::MallocSizeOf as DeriveMallocSizeOf;
-use metrics::{
-    register_meter_with_group, Gauge, GaugeUsize, Meter, MeterTimer,
-};
-use parking_lot::{Mutex, RwLock};
-use primitives::{
-    compute_block_number,
-    epoch::BlockHashOrEpochNumber,
-    filter::{FilterError, LogFilter},
-    log_entry::LocalizedLogEntry,
-    pos::PosBlockId,
-    receipt::Receipt,
-    Block, EpochId, EpochNumber, SignedTransaction, TransactionIndex,
-    TransactionStatus,
-};
-use rayon::prelude::*;
-use std::{
-    cmp::{max, min},
-    collections::HashSet,
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc,
-    },
-    thread::sleep,
-    time::Duration,
-};
+use metrics::{Gauge, GaugeUsize};
+use std::sync::Arc;
 
-lazy_static!{
-    static ref BEST_EPOCH_NUMBER: Arc<dyn Gauge<usize>> = GaugeUsize::register_with_group("graph_statistic", "best_epoch_number");
+lazy_static! {
+    static ref BEST_EPOCH_NUMBER: Arc<dyn Gauge<usize>> =
+        GaugeUsize::register_with_group("graph_statistic", "best_epoch_number");
 }
 
 #[derive(Default, Debug, DeriveMallocSizeOf)]
@@ -100,7 +25,6 @@ impl BestInformation {
 }
 
 impl ConsensusGraph {
-    
     pub fn best_info(&self) -> Arc<BestInformation> {
         self.best_info.read_recursive().clone()
     }
@@ -135,9 +59,8 @@ impl ConsensusGraph {
     pub fn best_block_hash(&self) -> H256 {
         self.best_info.read_recursive().best_block_hash
     }
-    
 
-       /// This function is called after a new block appended to the
+    /// This function is called after a new block appended to the
     /// ConsensusGraph. Because BestInformation is often queried outside. We
     /// store a version of best_info outside the inner to prevent keep
     /// getting inner locks.
