+++ crates/cfxcore/core/src/consensus/consensus_graph/mining_api.rs
@@ -1,88 +1,14 @@
 use super::ConsensusGraph;
 
-pub use crate::consensus::{
-    consensus_inner::{ConsensusGraphInner, ConsensusInnerConfig},
-    consensus_trait::SharedConsensusGraph,
-};
-use crate::{
-    block_data_manager::{
-        BlockDataManager, BlockExecutionResultWithEpoch, DataVersionTuple,
-    },
-    consensus::{
-        consensus_inner::{
-            consensus_executor::ConsensusExecutionConfiguration, StateBlameInfo,
-        },
-        pos_handler::PosVerifier,
-    },
-    errors::{invalid_params, invalid_params_check, Result as CoreResult},
-    pow::{PowComputer, ProofOfWorkConfig},
-    statistics::SharedStatistics,
-    transaction_pool::SharedTransactionPool,
-    verification::VerificationConfig,
-    NodeType, Notifications,
-};
-use cfx_execute_helper::{
-    estimation::{EstimateExt, EstimateRequest},
-    exec_tracer::{
-        recover_phantom_traces, ActionType, BlockExecTraces, LocalizedTrace,
-        TraceFilter,
-    },
-    phantom_tx::build_bloom_and_recover_phantom,
-};
-use cfx_executor::{
-    executive::ExecutionOutcome, spec::CommonParams, state::State,
-};
-use cfx_rpc_eth_types::EvmOverrides;
-use geth_tracer::GethTraceWithHash;
+use crate::consensus::consensus_inner::{ConsensusGraphInner, StateBlameInfo};
 
-use alloy_rpc_types_trace::geth::GethDebugTracingOptions;
-use cfx_internal_common::ChainIdParams;
-use cfx_parameters::{
-    consensus::*,
-    consensus_internal::REWARD_EPOCH_COUNT,
-    rpc::{
-        GAS_PRICE_BLOCK_SAMPLE_SIZE, GAS_PRICE_DEFAULT_VALUE,
-        GAS_PRICE_TRANSACTION_SAMPLE_SIZE,
-    },
-};
-use cfx_rpc_cfx_types::PhantomBlock;
-use cfx_statedb::StateDb;
-use cfx_storage::{
-    state::StateTrait, state_manager::StateManagerTrait, StorageState,
-};
-use cfx_types::{AddressWithSpace, AllChainID, Bloom, Space, H256, U256};
-use either::Either;
-use itertools::Itertools;
-use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
-use malloc_size_of_derive::MallocSizeOf as DeriveMallocSizeOf;
-use metrics::{
-    register_meter_with_group, Gauge, GaugeUsize, Meter, MeterTimer,
-};
-use parking_lot::{Mutex, RwLock};
-use primitives::{
-    compute_block_number,
-    epoch::BlockHashOrEpochNumber,
-    filter::{FilterError, LogFilter},
-    log_entry::LocalizedLogEntry,
-    pos::PosBlockId,
-    receipt::Receipt,
-    Block, EpochId, EpochNumber, SignedTransaction, TransactionIndex,
-    TransactionStatus,
-};
-use rayon::prelude::*;
-use std::{
-    cmp::{max, min},
-    collections::HashSet,
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc,
-    },
-    thread::sleep,
-    time::Duration,
-};
+use cfx_parameters::consensus::*;
+use cfx_types::{H256, U256};
+use primitives::pos::PosBlockId;
+
+use std::{thread::sleep, time::Duration};
 
 impl ConsensusGraph {
-
     /// Determine whether the next mined block should have adaptive weight or
     /// not
     pub fn check_mining_adaptive_block(
@@ -112,7 +38,7 @@ impl ConsensusGraph {
         )
     }
 
-        /// Wait for the generation and the execution completion of a block in the
+    /// Wait for the generation and the execution completion of a block in the
     /// consensus graph. This API is used mainly for testing purpose
     pub fn wait_for_generation(&self, hash: &H256) {
         while !self
@@ -143,7 +69,7 @@ impl ConsensusGraph {
         }
     }
 
-        /// After considering the latest `pos_reference`, `parent_hash` may become
+    /// After considering the latest `pos_reference`, `parent_hash` may become
     /// an invalid choice, so this function tries to update the parent and
     /// referee choices with `pos_reference` provided.
     pub fn choose_correct_parent(
@@ -223,7 +149,7 @@ impl ConsensusGraph {
         }
     }
 
-        /// Force the engine to recompute the deferred state root for a particular
+    /// Force the engine to recompute the deferred state root for a particular
     /// block given a delay.
     pub fn force_compute_blame_and_deferred_state_for_generation(
         &self, parent_block_hash: &H256,
