+++ crates/cfxcore/core/src/consensus/consensus_graph/mod.rs
@@ -1,92 +1,45 @@
-pub(super) mod onchain_blocks_provider;
 pub(super) mod best_info_provider;
-pub(super) mod sync_graph_api;
 pub(super) mod mining_api;
+pub(super) mod onchain_blocks_provider;
 pub(super) mod rpc_api;
+pub(super) mod sync_graph_api;
 
 use self::best_info_provider::BestInformation;
 
-use super::{consensus_inner::{
-    confirmation_meter::ConfirmationMeter,
-    consensus_executor::ConsensusExecutor,
-    consensus_new_block_handler::ConsensusNewBlockHandler,
-}, pivot_hint::PivotHint};
-pub use crate::consensus::{
-    consensus_inner::{ConsensusGraphInner, ConsensusInnerConfig},
-    consensus_trait::SharedConsensusGraph,
+use super::{
+    consensus_inner::{
+        confirmation_meter::ConfirmationMeter,
+        consensus_executor::ConsensusExecutor,
+        consensus_new_block_handler::ConsensusNewBlockHandler,
+    },
+    pivot_hint::PivotHint,
 };
+pub use crate::consensus::consensus_inner::ConsensusGraphInner;
 use crate::{
-    block_data_manager::{
-        BlockDataManager, BlockExecutionResultWithEpoch, DataVersionTuple,
-    }, consensus::{
-        consensus_inner::{
-            consensus_executor::ConsensusExecutionConfiguration, StateBlameInfo,
-        },
+    block_data_manager::BlockDataManager,
+    consensus::{
+        consensus_inner::consensus_executor::ConsensusExecutionConfiguration,
         pos_handler::PosVerifier,
-    }, errors::{invalid_params, invalid_params_check, Result as CoreResult}, pow::{PowComputer, ProofOfWorkConfig}, statistics::SharedStatistics, transaction_pool::SharedTransactionPool, verification::VerificationConfig, NodeType, Notifications
-};
-use cfx_execute_helper::{
-    estimation::{EstimateExt, EstimateRequest},
-    exec_tracer::{
-        recover_phantom_traces, ActionType, BlockExecTraces, LocalizedTrace,
-        TraceFilter,
     },
-    phantom_tx::build_bloom_and_recover_phantom,
+    pow::{PowComputer, ProofOfWorkConfig},
+    statistics::SharedStatistics,
+    transaction_pool::SharedTransactionPool,
+    verification::VerificationConfig,
+    NodeType, Notifications,
 };
-use cfx_executor::{
-    executive::ExecutionOutcome, spec::CommonParams, state::State,
-};
-use cfx_rpc_eth_types::EvmOverrides;
+
+use cfx_executor::spec::CommonParams;
+
 use super::config::ConsensusConfig;
-use geth_tracer::GethTraceWithHash;
 
-use alloy_rpc_types_trace::geth::GethDebugTracingOptions;
-use cfx_internal_common::ChainIdParams;
-use cfx_parameters::{
-    consensus::*,
-    consensus_internal::REWARD_EPOCH_COUNT,
-    rpc::{
-        GAS_PRICE_BLOCK_SAMPLE_SIZE, GAS_PRICE_DEFAULT_VALUE,
-        GAS_PRICE_TRANSACTION_SAMPLE_SIZE,
-    },
-};
-use cfx_rpc_cfx_types::PhantomBlock;
-use cfx_statedb::StateDb;
-use cfx_storage::{
-    state::StateTrait, state_manager::StateManagerTrait, StorageState,
-};
-use cfx_types::{AddressWithSpace, AllChainID, Bloom, Space, H256, U256};
-use either::Either;
-use itertools::Itertools;
+use cfx_types::H256;
+
 use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
-use malloc_size_of_derive::MallocSizeOf as DeriveMallocSizeOf;
-use metrics::{
-    register_meter_with_group, Gauge, GaugeUsize, Meter, MeterTimer,
-};
+
 use parking_lot::{Mutex, RwLock};
-use primitives::{
-    compute_block_number,
-    epoch::BlockHashOrEpochNumber,
-    filter::{FilterError, LogFilter},
-    log_entry::LocalizedLogEntry,
-    pos::PosBlockId,
-    receipt::Receipt,
-    Block, EpochId, EpochNumber, SignedTransaction, TransactionIndex,
-    TransactionStatus,
-};
-use rayon::prelude::*;
-use std::{
-    cmp::{max, min},
-    collections::HashSet,
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc,
-    },
-    thread::sleep,
-    time::Duration,
-};
+use primitives::EpochId;
 
-
+use std::sync::{atomic::AtomicBool, Arc};
 
 /// ConsensusGraph is a layer on top of SynchronizationGraph. A SyncGraph
 /// collect all blocks that the client has received so far, but a block can only
@@ -240,7 +193,7 @@ impl ConsensusGraph {
 
     /// Get the number of processed blocks (i.e., the number of calls to
     /// on_new_block()
-    #[cfg(feature="consensus_bench")]
+    #[cfg(feature = "consensus_bench")]
     pub fn get_processed_block_count(&self) -> usize {
         self.statistics.get_consensus_graph_processed_block_count()
     }
@@ -250,7 +203,6 @@ impl ConsensusGraph {
     pub fn data_manager(&self) -> &Arc<BlockDataManager> { &self.data_man }
 
     pub fn tx_pool(&self) -> &SharedTransactionPool { &self.txpool }
-
 }
 
 impl Drop for ConsensusGraph {
