+++ crates/cfxcore/core/src/consensus/consensus_graph/rpc_api/mod.rs
@@ -1,5 +1,5 @@
 pub mod execution_provider;
+pub mod phantom_block_provider;
 pub mod state_provider;
 pub mod trace_provider;
 pub mod transaction_provider;
-pub mod phantom_block_provider;
