+++ crates/cfxcore/core/src/consensus/consensus_graph/rpc_api/transaction_provider.rs
@@ -1,82 +1,6 @@
-pub use crate::consensus::{
-    consensus_inner::{ConsensusGraphInner, ConsensusInnerConfig},
-    consensus_trait::SharedConsensusGraph,
-};
-use crate::{
-    block_data_manager::{
-        BlockDataManager, BlockExecutionResultWithEpoch, DataVersionTuple,
-    },
-    consensus::{
-        consensus_inner::{
-            consensus_executor::ConsensusExecutionConfiguration, StateBlameInfo,
-        },
-        pos_handler::PosVerifier,
-    },
-    errors::{invalid_params, invalid_params_check, Result as CoreResult},
-    pow::{PowComputer, ProofOfWorkConfig},
-    statistics::SharedStatistics,
-    transaction_pool::SharedTransactionPool,
-    verification::VerificationConfig,
-    NodeType, Notifications,
-};
-use cfx_execute_helper::{
-    estimation::{EstimateExt, EstimateRequest},
-    exec_tracer::{
-        recover_phantom_traces, ActionType, BlockExecTraces, LocalizedTrace,
-        TraceFilter,
-    },
-    phantom_tx::build_bloom_and_recover_phantom,
-};
-use cfx_executor::{
-    executive::ExecutionOutcome, spec::CommonParams, state::State,
-};
-use cfx_rpc_eth_types::EvmOverrides;
-use geth_tracer::GethTraceWithHash;
-
-use alloy_rpc_types_trace::geth::GethDebugTracingOptions;
-use cfx_internal_common::ChainIdParams;
-use cfx_parameters::{
-    consensus::*,
-    consensus_internal::REWARD_EPOCH_COUNT,
-    rpc::{
-        GAS_PRICE_BLOCK_SAMPLE_SIZE, GAS_PRICE_DEFAULT_VALUE,
-        GAS_PRICE_TRANSACTION_SAMPLE_SIZE,
-    },
-};
-use cfx_rpc_cfx_types::PhantomBlock;
-use cfx_statedb::StateDb;
-use cfx_storage::{
-    state::StateTrait, state_manager::StateManagerTrait, StorageState,
-};
-use cfx_types::{AddressWithSpace, AllChainID, Bloom, Space, H256, U256};
-use either::Either;
-use itertools::Itertools;
-use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
-use malloc_size_of_derive::MallocSizeOf as DeriveMallocSizeOf;
-use metrics::{
-    register_meter_with_group, Gauge, GaugeUsize, Meter, MeterTimer,
-};
-use parking_lot::{Mutex, RwLock};
+use cfx_types::{H256, U256};
 use primitives::{
-    compute_block_number,
-    epoch::BlockHashOrEpochNumber,
-    filter::{FilterError, LogFilter},
-    log_entry::LocalizedLogEntry,
-    pos::PosBlockId,
-    receipt::Receipt,
-    Block, EpochId, EpochNumber, SignedTransaction, TransactionIndex,
-    TransactionStatus,
-};
-use rayon::prelude::*;
-use std::{
-    cmp::{max, min},
-    collections::HashSet,
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc,
-    },
-    thread::sleep,
-    time::Duration,
+    receipt::Receipt, SignedTransaction, TransactionIndex, TransactionStatus,
 };
 
 use super::super::ConsensusGraph;
@@ -99,29 +23,25 @@ impl ConsensusGraph {
     ) -> Option<(SignedTransaction, TransactionInfo)> {
         // We need to hold the inner lock to ensure that tx_index and receipts
         // are consistent
-        
-        let tx_info = self.get_transaction_info(hash)?; 
-            if let Some(executed) = &tx_info.maybe_executed_extra_info {
-                if executed.receipt.outcome_status == TransactionStatus::Skipped
-                {
-                    // A skipped transaction is not visible to clients if
-                    // accessed by its hash.
-                    return None;
-                }
+
+        let tx_info = self.get_transaction_info(hash)?;
+        if let Some(executed) = &tx_info.maybe_executed_extra_info {
+            if executed.receipt.outcome_status == TransactionStatus::Skipped {
+                // A skipped transaction is not visible to clients if
+                // accessed by its hash.
+                return None;
             }
-            let block = self.data_man.block_by_hash(
-                &tx_info.tx_index.block_hash,
-                false, /* update_cache */
-            )?;
-            let transaction =
-                (*block.transactions[tx_info.tx_index.real_index]).clone();
-            Some((transaction, tx_info))
-      
+        }
+        let block = self.data_man.block_by_hash(
+            &tx_info.tx_index.block_hash,
+            false, /* update_cache */
+        )?;
+        let transaction =
+            (*block.transactions[tx_info.tx_index.real_index]).clone();
+        Some((transaction, tx_info))
     }
 
-    fn get_transaction_info(
-        &self, tx_hash: &H256,
-    ) -> Option<TransactionInfo> {
+    fn get_transaction_info(&self, tx_hash: &H256) -> Option<TransactionInfo> {
         let inner = self.inner.read();
 
         trace!("Get receipt with tx_hash {}", tx_hash);
