+++ crates/stratum/src/lib.rs
@@ -266,9 +266,7 @@ impl Default for SocketMetadata {
 }
 
 impl SocketMetadata {
-    pub fn addr(&self) -> &SocketAddr {
-        &self.addr
-    }
+    pub fn addr(&self) -> &SocketAddr { &self.addr }
 }
 
 impl Metadata for SocketMetadata {}
@@ -306,9 +304,7 @@ mod tests {
     pub struct VoidManager;
 
     impl JobDispatcher for VoidManager {
-        fn submit(&self, _payload: Vec<String>) -> Result<(), Error> {
-            Ok(())
-        }
+        fn submit(&self, _payload: Vec<String>) -> Result<(), Error> { Ok(()) }
     }
 
     fn dummy_request(addr: &SocketAddr, data: &str) -> Vec<u8> {
@@ -368,9 +364,7 @@ mod tests {
     }
 
     impl JobDispatcher for DummyManager {
-        fn submit(&self, _payload: Vec<String>) -> Result<(), Error> {
-            Ok(())
-        }
+        fn submit(&self, _payload: Vec<String>) -> Result<(), Error> { Ok(()) }
     }
 
     fn terminated_str(origin: &'static str) -> String {
